+++ Stidents.BLL/Services/ManagerService.cs
@@ -1,24 +1,23 @@
-﻿using Microsoft.Extensions.Caching.Memory;
-using Microsoft.Extensions.Logging;
+﻿using Microsoft.Extensions.Logging;
 using Students.BLL.DataAccess;
 using Students.DAL.Models;
 using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Students.DAL.Enum;
+using System.Linq.Dynamic.Core;
+using System.Linq;
 
 namespace Students.BLL.Services
 {
     public class ManagerService : IManagerService
     {
         private readonly UnitOfWork _unitOfWork;
-        private readonly IMemoryCache cache;
         private readonly ILogger _logger;
 
-        public ManagerService(UnitOfWork unitOfWork, IMemoryCache memoryCache, ILogger<Manager> logger)
+        public ManagerService(UnitOfWork unitOfWork, ILogger<Manager> logger)
         {
             _unitOfWork = unitOfWork;
-            cache = memoryCache;
             _logger = logger;
         }
 
@@ -27,6 +26,7 @@ namespace Students.BLL.Services
             try
             {
                 await _unitOfWork.ManagerRepository.CreateAsync(item);
+                await _unitOfWork.SaveAsync();
                 _logger.LogInformation("Менеджер создан");
             }
             catch (Exception ex)
@@ -39,8 +39,14 @@ namespace Students.BLL.Services
         {
             try
             {
-                await _unitOfWork.ManagerRepository.DeleteAsync(id); ;
-                _logger.LogInformation(id, "Менеджер удален"); ;
+                Manager manager = await GetAsync(id);
+                if (manager != null)
+                {
+                    await _unitOfWork.ApplicationUsers.DeleteAsync(manager.UserId);
+                    await _unitOfWork.ManagerRepository.DeleteAsync(id);
+                    await _unitOfWork.SaveAsync();
+                    _logger.LogInformation(id, "Менеджер удален"); ;
+                }
             }
             catch (Exception ex)
             {
@@ -63,7 +69,7 @@ namespace Students.BLL.Services
             catch (Exception ex)
             {
                 _logger.LogError(ex, "Ошибка получение списка менеджеров");
-                return null;
+                return Enumerable.Empty<Manager>();
             }
         }
 
@@ -86,6 +92,7 @@ namespace Students.BLL.Services
             try
             {
                 var manager = await _unitOfWork.ManagerRepository.Update(item);
+                await _unitOfWork.SaveAsync();
                 _logger.LogInformation("Менеджер изменен");
                 return manager;
             }
@@ -96,17 +103,12 @@ namespace Students.BLL.Services
             }
         }
 
-        public async Task<IEnumerable<Manager>>  GetAllTakeSkipAsync(int take, EnumPageActions action, int skip = 0)
-        {
-            return await _unitOfWork.ManagerRepository.GetAllTakeSkipAsync(take, action, skip);
-        }
-
-        public async Task<Manager> SearchAsync(string predicate)
+        public async Task<Manager> SearchAsync(string query)
         {
             try
             {
                 _logger.LogInformation("Поиск менаджера");
-                return await _unitOfWork.ManagerRepository.SearchAsync(predicate);
+                return await _unitOfWork.ManagerRepository.SearchAsync(query);
             }
             catch (Exception ex)
             {
@@ -115,19 +117,46 @@ namespace Students.BLL.Services
             }
         }
 
-        public async Task<IEnumerable<Manager>> SearchAllAsync(string searchString, EnumSearchParameters searchParameter, EnumPageActions action, int take, int skip = 0)
+        public async Task<int> GetCount(string searchString, EnumSearchParameters searchParametr)
         {
-            return await _unitOfWork.ManagerRepository.SearchAllAsync(searchString,searchParameter,action, take, skip);
+            if (string.IsNullOrEmpty(searchString) || searchParametr == EnumSearchParameters.None)
+            {
+                return (await _unitOfWork.ManagerRepository.GetAllAsync()).Count();
+            }
+            return (await SearchAllAsync(searchString, searchParametr)).Count();
         }
 
-        public async Task<IEnumerable<Manager>> DisplayingIndex(EnumPageActions action, string searchString, EnumSearchParameters searchParametr, int take, int skip = 0)
+        public async Task<IEnumerable<Manager>> GetPaginatedResult(int currentPage, int pageSize = 10)
         {
-            take = (take == 0) ? 10 : take;
-            if (!String.IsNullOrEmpty(searchString))
+            return (await _unitOfWork.ManagerRepository.GetAllAsync())
+                .OrderBy(t => t.Surname).Skip((currentPage - 1) * pageSize).Take(pageSize);
+        }
+
+        public async Task<IEnumerable<Manager>> SearchAllAsync(string searchString, EnumSearchParameters searchParametr)
+        {
+            if (string.IsNullOrEmpty(searchString) || searchParametr == EnumSearchParameters.None)
+                return Enumerable.Empty<Manager>();
+            return (await _unitOfWork.ManagerRepository.GetAllAsync()).AsQueryable()
+                .Where($"{searchParametr.ToString().Replace('_', '.')}.Contains(@0)", searchString);
+        }
+
+        public async Task<IEnumerable<Manager>> SearchAllAsync(string searchString, EnumSearchParameters searchParametr, int currentPage, int pageSize)
+        {
+            if (string.IsNullOrEmpty(searchString) || searchParametr == EnumSearchParameters.None)
+                return Enumerable.Empty<Manager>();
+            return (await _unitOfWork.ManagerRepository.GetAllAsync()).AsQueryable()
+                .OrderBy(t => t.Surname)
+                .Where($"{searchParametr.ToString().Replace('_', '.')}.Contains(@0)", searchString)
+                .Skip((currentPage - 1) * pageSize).Take(pageSize);
+        }
+
+        public async Task<IEnumerable<Manager>> IndexView(string searchString, EnumSearchParameters searchParametr, int currentPage, int pageSize = 10)
+        {
+            if (!String.IsNullOrEmpty(searchString) && searchParametr != EnumSearchParameters.None)
             {
-                return await SearchAllAsync(searchString, searchParametr, action, take, skip);
+                return await SearchAllAsync(searchString, searchParametr, currentPage, pageSize);
             }
-            return await GetAllTakeSkipAsync(take, action, skip);
+            return await GetPaginatedResult(currentPage, pageSize);
         }
     }
 }
